main function for parameter estimation

main function for parameter estimation

main_optim(obs_list, crit_function, model_function, model_options = NULL,
  optim_method = "simplex", optim_options = NULL, prior_information)

Arguments

obs_list
List of observed values

crit_function
Function implementing the criterion to optimize

model_function
Crop Model wrapper function

model_options
List of options for the Crop Model wrapper (optional, see help of the Crop Model wrapper used)

optim_method
Name of the parameter estimation method (optional, see help of optim_switch function)

optim_options
List of options of the parameter estimation method: nb_rep, the number of repetitions (optional, default=1) (see help of optim_switch function))

prior_information
Prior information on the parameters to estimate. For the moment only uniform distribution are allowed. Either a list containing (named) vectors of upper and lower bounds (ub and lb), or a named list containing for each parameter the list of situations per group (sit_list) and the vector of upper and lower bounds (one value per group) (ub and lb)

param_names
Name(s) of parameters to estimate

Value

The vector of values for optimized parameters + prints and graphs, depending on the parameter estimation method used

Examples

library(SticsRFiles)

#> Learn SticsRFiles at: https://SticsRPacks.github.io/SticsRFiles

#> 
#> Attachement du package : ‘SticsRFiles’

#> The following object is masked from ‘package:SticsOptimizR’:
#> 
#>     get_params_names

library(SticsOnR)

#> Registered S3 method overwritten by 'sensitivity':
#>   method    from 
#>   print.src dplyr

#> Learn SticsOnR at: https://SticsRPacks.github.io/SticsOnR

#> 
#> Attachement du package : ‘SticsOnR’

#> The following objects are masked from ‘package:SticsRFiles’:
#> 
#>     set_file, set_general, set_ini, set_out_var, set_param, set_plant,
#>     set_soil, set_station, set_tec, set_tmp, set_usm

library(SticsOptimizR)
library(dplyr)

#> 
#> Attachement du package : ‘dplyr’

#> The following object is masked from ‘package:SticsRFiles’:
#> 
#>     rename

#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag

#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union

library(nloptr)
library(DiceDesign)
library("doParallel")

#> Le chargement a nécessité le package : foreach

#> Le chargement a nécessité le package : iterators

#> Le chargement a nécessité le package : parallel

# TEST ON ONE SITUATION and ONE VARIABLE

# Select situation and variable
sit_name="bo96iN+"
var_name="lai_n"

# Run the model before optimization for a prior evaluation
stics_path="D:/Home/sbuis/Documents/WORK/STICS/JavaSTICS-1.41-stics-9.0/bin/stics_modulo"  # TO ADAPT TO YOUR CASE :-) ###

data_dir=system.file(file.path("extdata","TestCase1c"), package = "SticsOptimizR")
model_options=stics_wrapper_options(stics_path,data_dir,parallel=FALSE)
sim_before_optim=stics_wrapper(model_options=model_options)

#> Error in stics_wrapper(model_options = model_options): Stics executable file doesn't exist ! D:/Home/sbuis/Documents/WORK/STICS/JavaSTICS-1.41-stics-9.0/bin/stics_modulo

# Read and select the corresponding observations
obs_list=read_obs_to_list(file.path(model_options$data_dir,"Orig Data"),
                          obs_filenames = paste0(sit_name,".obs"))

#> Warning: NAs introduits lors de la conversion automatique

obs_list[[sit_name]]=obs_list[[sit_name]][,c("Date",var_name)]

# Set prior information on the parameters to estimate
prior_information=list(lb=c(dlaimax=0.0005, durvieF=50),
                       ub=c(dlaimax=0.0025, durvieF=400))

# Set options for the parameter estimation method
optim_options=list()
optim_options$nb_rep <- 2 # How many times we run the minimization with different parameters
optim_options$xtol_rel <- 1e-05 # Tolerance criterion between two iterations
optim_options$maxeval <- 20 # Maximum number of iterations executed by the function
optim_options$path_results <- model_options$data_dir # path where to store results graphs

# Run the optimization
param_est_values=main_optim(obs_list=obs_list,crit_function=concentrated_wss,
                            model_function=stics_wrapper,
                            model_options=model_options,
                            optim_options=optim_options,
                            prior_information=prior_information)

#> Error in model_function(param_values = param_values, sit_var_dates_mask = obs_list,     prior_information = prior_information, model_options = model_options): Stics executable file doesn't exist ! D:/Home/sbuis/Documents/WORK/STICS/JavaSTICS-1.41-stics-9.0/bin/stics_modulo

# Run the model after optimzation
sim_after_optim=stics_wrapper(param_values=param_est_values,model_options=model_options)

#> Error in stics_wrapper(param_values = param_est_values, model_options = model_options): Stics executable file doesn't exist ! D:/Home/sbuis/Documents/WORK/STICS/JavaSTICS-1.41-stics-9.0/bin/stics_modulo

# Plot the results

dev.new()

#> dev.new(): using pdf(file="Rplots24.pdf")

par(mfrow = c(1,2))
Ymax=max(max(obs_list[[sit_name]][,var_name], na.rm=TRUE),
         max(sim_before_optim$sim_list[[sit_name]][,var_name], na.rm=TRUE))

#> Error in eval(expr, envir, enclos): objet 'sim_before_optim' introuvable

plot(sim_before_optim$sim_list[[sit_name]][,c("Date",var_name)],type="l",
     main="Before optimization",ylim=c(0,Ymax+Ymax*0.1))

#> Error in plot(sim_before_optim$sim_list[[sit_name]][, c("Date", var_name)],     type = "l", main = "Before optimization", ylim = c(0, Ymax +         Ymax * 0.1)): objet 'sim_before_optim' introuvable

points(obs_list[[sit_name]],col="green")

#> Error in plot.xy(xy.coords(x, y), type = type, ...): plot.new has not been called yet

plot(sim_after_optim$sim_list[[sit_name]][,c("Date",var_name)],type="l",
     main="After optimization",ylim=c(0,Ymax+Ymax*0.1))

#> Error in plot(sim_after_optim$sim_list[[sit_name]][, c("Date", var_name)],     type = "l", main = "After optimization", ylim = c(0, Ymax +         Ymax * 0.1)): objet 'sim_after_optim' introuvable

points(obs_list[[sit_name]],col="green")

#> Error in plot.xy(xy.coords(x, y), type = type, ...): plot.new has not been called yet

# TEST WITH SEVERAL SITUATIONS AND USING GROUPS OF SITUATIONS PER PARAMETER

prior_information=list()
prior_information$dlaimax=list(sit_list=list(c("bou99t3", "bou00t3", "bou99t1", "bou00t1", "bo96iN+", "lu96iN+", "lu96iN6", "lu97iN+")),lb=0.0005,ub=0.0025)
prior_information$durvieF=list(sit_list=list(c("bo96iN+", "lu96iN+", "lu96iN6", "lu97iN+"), c("bou99t3", "bou00t3", "bou99t1", "bou00t1")),lb=c(50,50),ub=c(400,400))

stics_path="D:/Home/sbuis/Documents/WORK/STICS/JavaSTICS-1.41-stics-9.0/bin/stics_modulo"  # TO ADAPT TO YOUR CASE :-) ###
data_dir=system.file(file.path("extdata","TestCase1c"), package = "SticsOptimizR")
model_options=stics_wrapper_options(stics_path,data_dir, parallel=TRUE)

optim_options=list()
optim_options$nb_rep=2                              # How many times we run the minimization with different parameters
optim_options$xtol_rel=1e-05                        # Tolerance criterion between two iterations
optim_options$maxeval=3                             # Maximum number of iterations executed by the function
optim_options$path_results=model_options$data_dir   # path where to store results graphs

obs_list=read_obs_to_list(file.path(model_options$data_dir,"Orig Data"))

#> Warning: NAs introduits lors de la conversion automatique

#> Warning: NAs introduits lors de la conversion automatique

#> Warning: NAs introduits lors de la conversion automatique

#> Warning: NAs introduits lors de la conversion automatique

#> Warning: NAs introduits lors de la conversion automatique

#> Warning: NAs introduits lors de la conversion automatique

#> Warning: NAs introduits lors de la conversion automatique

#> Warning: NAs introduits lors de la conversion automatique

main_optim(obs_list=obs_list,crit_function=concentrated_wss,model_function=stics_wrapper,model_options=model_options,optim_options=optim_options,prior_information=prior_information)

#> Error in model_function(param_values = param_values, sit_var_dates_mask = obs_list,     prior_information = prior_information, model_options = model_options): Stics executable file doesn't exist ! D:/Home/sbuis/Documents/WORK/STICS/JavaSTICS-1.41-stics-9.0/bin/stics_modulo